requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
h5py==3.1.0
numpy==1.19.5
pandas==1.3.1
pymongo==3.11.2
regex==2021.8.28
scikit-learn==0.24.2
spacy==3.1.3
tensorflow==2.5.0
torch==1.9.1
transformers==4.10.2